pyproject.toml
[tool.poetry]
name = "ccdb"
version = "0.1.0"
description = ""
authors = ["Sören Berger <soeren.berger@dc1.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9.1"
celery = "5.3.1"
django = "^4.2.0"
django-crum = "^0.7.2"
django-csp = "^3.5"
django-extensions = "^3.2.3"
django-filter = "^2.0.0"
django-import-export = "^4.3.5"
django-jsoneditor = "^0.2.2"
django-lifecycle = "^1.2.4"
django-model-utils = "^5.0"
django-simple-history = "^3.8.0"
djangorestframework = "^3.8.2"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.8.5"
pylibmc = "^1.5.2"
python-dateutil = "^2.7.3"
requests = "^2.24.0"
tqdm = "^4.63.1"
coreapi = "^2.3.3"
pika = "^1.3.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.4.3"
django-coverage-plugin = "^3.1.0"
django-test-without-migrations = "^0.6"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
pytest-custom-exit-code = "^0.3.0"
pytest-django = "^4.8.0"
pytest-env = "^1.1.3"
pytest-rerunfailures = "^13.0"
pytest-sugar = "^1.0.0"
pytest-xdist = "^3.3.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
djhtml = "^3.0.6"
flake8 = "^7.0.0"
flake8-bugbear = "^24.2.6"
ipdb = "^0.13.7"
ipython = "^7.22"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=false
use_parentheses=true
line_length=88
skip = ["migrations","settings.py","wsgi.py","celery_app.py","test_settings.py",".tox","local"]
known_first_party = ["api", "ccdb", "contracting", "easybill", "export", "globalways", "gwdb", "main"]
known_third_party = ["rest_framework"]

[tool.coverage.run]
branch = true
omit = [
  "*/admin.py",
  "*/local/*",
  "*/migrations/*",
  "*/settings*",
  "*/settings*",
  "*/urls.py",
  "*/manage.py",
  "*/setup.py",
  "*/wsgi.py",
  "*/__init__.py",
  "*/__main__.py",
  "*/celery_app.py",
]

[tool.coverage.report]
exclude_also = [
  "pragma: no cover",
  "def __str__",
  "der __repr__",
  "if settings.DEBUG",
  "NOQA",
  "NotImplementedError",
]

[tool.flake8]
ignore = ["E203", "E231", "E266", "E501", "W503", "W605"]
max-line-length = 160
exclude = [
  "migrations",
  "static",
  "_static",
  "build",
  "settings.py",
  "tox/*",
  "local",
  "__init__.py",
  "test_*",
]

[tool.pytest.ini_options]
addopts = "--nomigrations --reuse-db --cov --cov-report term-missing:skip-covered"
env = "DJANGO_SETTINGS_MODULE=ccdb.settings.citest"
